Split config module
