+++ src/config/mod.rs
@@ -1,0 +1,9 @@
+pub mod config;
+pub mod error;
+pub mod gobjects;
+pub mod regex_list;
+pub mod work_mode;
+
+pub use self::config::Config;
+pub use self::regex_list::RegexList;
+pub use self::work_mode::WorkMode;
